music refactor
music player and music manager
