+++ src/main/java/Controlers/PlayController.java
@@ -1,6 +1,7 @@
 package Controlers;
 
 import Services.*;
+import Services.Music.MusicManager;
 import Shapes.Block;
 import Tetris.Game;
 import Tetris.ShapeFactory;
@@ -47,7 +48,7 @@ public class PlayController extends Controller implements Initializable, Observe
     @Override
     public void initialize(final URL location, final ResourceBundle resources) {
         ControllerManager.setPlayController(this);
-        MusicPlayer.INSTANCE.bindAudioButtonImage(soundButton);
+        MusicManager.INSTANCE.bindAudioButtonImage(soundButton);
         Platform.runLater(this::initGame);
         graphics = new Graphics(canvasForBigPane, canvasForSmallPane);
         graphics.setUpGraphics(bigPane, smallPane);
