+++ src/main/java/Tetris/Main.java
@@ -1,18 +1,14 @@
 package Tetris;
 
-import Services.MusicPlayer;
+import Services.Music.MusicManager;
 import javafx.application.Application;
 import javafx.application.Platform;
-import javafx.event.EventHandler;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.scene.text.Font;
-import javafx.stage.Modality;
 import javafx.stage.Stage;
-import javafx.stage.StageStyle;
-import javafx.stage.WindowEvent;
 
 import java.io.InputStream;
 
@@ -24,7 +20,7 @@ public class Main extends Application {
         primaryStage.setScene(new Scene(root,550,850));
         primaryStage.setTitle("Gran Canaria Tetris - version 1.0");
 
-        MusicPlayer.INSTANCE.play();
+        MusicManager.INSTANCE.initMusic();
 
         InputStream inputStream = this.getClass().getResourceAsStream("/images/horse_icon.jpg");
         primaryStage.getIcons().add(new Image(inputStream));
